+++ ext/stub.c
@@ -240,6 +240,7 @@ zend_class_entry *stub_issue2682_ce;
 zend_class_entry *stub_issue2691_ce;
 zend_class_entry *stub_issue2698_ce;
 zend_class_entry *stub_issue2699_ce;
+zend_class_entry *stub_issue2702_ce;
 zend_class_entry *stub_issue663_ce;
 zend_class_entry *stub_issue760_ce;
 zend_class_entry *stub_issue808_ce;
@@ -552,6 +553,7 @@ static PHP_MINIT_FUNCTION(stub)
 	ZEPHIR_INIT(Stub_Issue2691);
 	ZEPHIR_INIT(Stub_Issue2698);
 	ZEPHIR_INIT(Stub_Issue2699);
+	ZEPHIR_INIT(Stub_Issue2702);
 	ZEPHIR_INIT(Stub_Issue663);
 	ZEPHIR_INIT(Stub_Issue760);
 	ZEPHIR_INIT(Stub_Issue808);